Separate type declarations, classes into separate files
